requirements.txt
fastapi>=0.111,<1.0
google-genai>=1.29,<2.0
httpx>=0.27,<1.0
pydantic>=2.5,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
tenacity>=8.2,<10.0
uvicorn[standard]>=0.30,<1.0
